General improvements (#3)
* Update cvpr-2024.py file

* Replace old repository name and add .block_until_ready() command

* Make adjustments to sphinx docs

* Add 3D shepp logan demo to demo folder

* Include slider display and reorganize plot utilities.

* Include slider display and reorganize plot utilities.

* Refactored forward_model_loss()

* Refactor get_estimate_of_recon_std()

* Refactor gen_set_of_voxel_partitions()

---------
